Fix 32-bit compile warning in sprintflst.c

+++ src/sprintflst.c
@@ -927,8 +927,8 @@ int rig_sprintf_agc_levels(RIG *rig, char *str, int lenstr)
             }
             else
             {
-                rig_debug(RIG_DEBUG_ERR, "%s: buffer overrun!!  len=%ld > maxlen=%d\n",
-                          __func__, strlen(str) + strlen(tmpbuf), lenstr - 1);
+                rig_debug(RIG_DEBUG_ERR, "%s: buffer overrun!!  len=%d > maxlen=%d\n",
+                          __func__, (int)(strlen(str) + strlen(tmpbuf)), lenstr - 1);
             }
         }
     }
@@ -947,8 +947,8 @@ int rig_sprintf_agc_levels(RIG *rig, char *str, int lenstr)
             }
             else
             {
-                rig_debug(RIG_DEBUG_ERR, "%s: buffer overrun!!  len=%ld > maxlen=%d\n",
-                          __func__, strlen(str) + strlen(tmpbuf), lenstr - 1);
+                rig_debug(RIG_DEBUG_ERR, "%s: buffer overrun!!  len=%d > maxlen=%d\n",
+                          __func__, (int)(strlen(str) + strlen(tmpbuf)), lenstr - 1);
             }
         }
     }
